requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
tqdm
pytest
pytest-mock
pytest-cov
pre-commit
black
isort
flake8
